[MIG] website_sale_require_login: Migration to 17.0
