+++ website_sale_require_login/__manifest__.py
@@ -3,13 +3,18 @@
 {
     "name": "Require login to checkout",
     "summary": "Force users to login for buying",
-    "version": "14.0.1.0.0",
+    "version": "17.0.1.0.0",
     "category": "Website",
     "website": "https://github.com/OCA/e-commerce",
     "author": "Tecnativa, " "LasLabs, " "Odoo Community Association (OCA)",
     "license": "AGPL-3",
     "installable": True,
     "depends": ["website_sale_suggest_create_account", "web_tour"],
-    "data": ["views/website_sale.xml", "views/assets.xml"],
+    "data": ["views/website_sale.xml"],
+    "assets": {
+        "web.assets_tests": [
+            "website_sale_require_login/static/tests/tours/checkout.esm.js",
+        ],
+    },
     "post_init_hook": "post_init_hook",
 }
