--- website_sale_require_login/views/assets.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<odoo>
-    <template
-        id="assets_tests"
-        name="Website Sale Require Login - Shop Buy Test"
-        inherit_id="web.assets_tests"
-    >
-        <xpath expr="." position="inside">
-            <script
-                type="text/javascript"
-                src="/website_sale_require_login/static/tests/tours/checkout.js"
-            />
-        </xpath>
-    </template>
-</odoo>
