+++ website_sale_require_login/views/website_sale.xml
@@ -8,32 +8,11 @@
         priority="20"
         customize_show="True"
     >
-        <xpath expr="//div[@id='wrap']" position="before">
+        <xpath expr="//t[@t-set='show_shorter_cart_summary']" position="before">
             <t
                 t-set="hide_checkout_button"
                 t-value="not (user_authenticated and can_checkout)"
             />
-        </xpath>
-        <xpath
-            expr="//div[hasclass('oe_cart')]//a[@href='/shop/checkout?express=1']"
-            position="attributes"
-        >
-            <attribute name="t-if">not hide_checkout_button</attribute>
-        </xpath>
-        <xpath
-            expr="//div[hasclass('col-12','col-xl-4')]//a[@href='/shop/checkout?express=1']"
-            position="attributes"
-        >
-            <attribute name="t-if">not hide_checkout_button</attribute>
-        </xpath>
-    </template>
-    <template
-        id="short_cart_summary"
-        inherit_id="website_sale.short_cart_summary"
-        customize_show="True"
-    >
-        <xpath expr="//a[@href='/shop/checkout?express=1']" position="attributes">
-            <attribute name="t-if">not hide_checkout_button</attribute>
         </xpath>
     </template>
 </odoo>
